chore(toasts): change to ts from jsx
